### 2023-01-20 #4935 @julianreul
@sbenthall Thank you for handling our submission. - If we can assist you during the review process, e.g. by proposing potential reviewers, just let us know. We're looking forward to publishing with JOSS.
Best regards
Julian
