### 2023-01-25 #4935 @julianreul
Thank you @sbenthall for taking over our submission as editor.

Based on the JOSS-reviewers spreadsheet I would propose **Christina Maimone** (E-Mail: christina.maimone@northwestern.edu, Username: cmaimone) as a potential reviewer. Her field of expertise (political science & social science, among others) seems to match our submission.

My second reviewer proposal would come from my personal network. However, I need to find out about his current workload first.

Best regards
Julian
