### 2023-01-26 aab4e51
MO|DE.behave V2.0.0: Introducing an updated algorithm for estimating MXL models.
