+++ mode_behave_public/post_analysis.py
@@ -10,12 +10,11 @@ to analyze the estimated mixed logit model.
 
 import pandas as pd
 import numpy as np
-from math import sqrt, pow, floor, exp
+from math import exp
 import matplotlib.pyplot as plt
 from matplotlib.lines import Line2D
 import matplotlib.patches as mpatches
 from sklearn.cluster import KMeans, AgglomerativeClustering, MeanShift, DBSCAN
-from operator import mod
 from numba import guvectorize, prange, njit
 import seaborn as sns
 
@@ -29,6 +28,17 @@ class PostAnalysis:
         pass       
     
     def get_AIC_MNL(self):
+        """
+        This method calculates the AIC of an estimated MNL-model.
+        The AIC is a measure to evaluate the quality of the estimated model.
+
+        Returns
+        -------
+        AIC : float
+            Akaike information criterion.
+
+        """
+        
         LL = self.loglike_MNL()
         
         AIC = -2*LL[0]+2*len(self.initial_point)
@@ -36,6 +46,17 @@ class PostAnalysis:
         return AIC
     
     def get_BIC_MNL(self):
+        """
+        This method calculates the BIC of an estimated MNL-model.
+        The BIC is a measure to evaluate the quality of the estimated model.
+
+        Returns
+        -------
+        BIC : float
+            Bayesian information criterion.
+
+        """
+        
         LL = self.loglike_MNL()
         
         BIC = -2*LL[0]+np.log(len(self.data))*len(self.initial_point)
@@ -53,7 +74,11 @@ class PostAnalysis:
 
         Returns
         -------
-        Probability of MNL model at a specified point.
+        res : float
+            Log-likelihood: Log-Probability of MNL model at a specified point.
+        number_nan : int
+            Number of occured nan-values: Should be zero, otherwise,
+            some numerical issue occures during calculation (e.g., log(0))
 
         """
         
@@ -144,6 +169,8 @@ class PostAnalysis:
     
     def loglike_MXL(self, **kwargs):
         """
+        This method calculates the log-likelihood for MXL-models.
+        
         For reference on log-likelihood calculation see:
             Ch. 5.5 (pp. 118) in "Discrete Choice Analysis", by Ben-Akiva (1985)
         
@@ -151,8 +178,6 @@ class PostAnalysis:
         -------
         Float64
             Returns the log-likelihood (LL) of the estimated mixed logit model.
-            The following should hold for model validity:
-                LL_mixed_logit > LL_initial_point (multinomial logit)
                 
         """
         points_in = kwargs.get('points_in', self.points)
@@ -163,6 +188,8 @@ class PostAnalysis:
             points_in=points_in, 
             vector_output_no_weights=True
             )
+        
+        self.logit_vector_check = logit_vector.copy()
         
         logit_vector_s = np.swapaxes(logit_vector, 0, 1)
         
@@ -180,6 +207,7 @@ class PostAnalysis:
         """
         This method returns the index of a given attribute name for the 
         array "self.initial_point".
+        This method is a supportive method for visualize_attribute()
 
         Parameters
         ----------
@@ -305,7 +333,6 @@ class PostAnalysis:
             else:
                 tick_label_temp.append("choice_" + str(i))
             
-        #continue with something as below!
         fig, ax = plt.subplots()
         title_temp = "Attribute: " + attribute
         ax.set_title(title_temp)
@@ -531,7 +558,7 @@ class PostAnalysis:
             cluster_labels_pd = cluster_labels_pd.reset_index(drop=True)
             cluster_labels_pd['weights'] = self.shares.values[index_clustered]
         else:
-            cluster_labels_pd['weights'] = self.shares.values
[…]
