+++ README.rst
@@ -6,11 +6,9 @@ the rapid quantitative analysis of survey data on choice behavior,
 utilizing advanced discrete choice methods. 
 Therefore, MO|DE.behave incorporates estimation routines for conventional 
 multinomial logit models, as well as for mixed logit models with nonparametric 
-distributions. As the estimation of this type of 
-mixed logit model can be computationally-expensive, the software facilitates the 
-use of GPU hardware during the estimation process in order to decrease computation time. 
+distributions.
 Furthermore, MO|DE.behave contains a set of post-processing tools for visualizing 
-estimation and simulated results. Additionally, pre-estimated 
+estimation and simulation results. Additionally, pre-estimated 
 discrete choice simulation methods for transportation research are included to 
 enrich the software package for this specific community.
 
@@ -37,8 +35,6 @@ Installation
     
       pip install -e .
       
-#. If a NVIDIA-GPU is accessible, gpu-based estimation methods can be used. Therefore, please also install the python-package "cuda" and an appropriate NVIDIA-driver.
-
 
 Workflow
 ========
@@ -48,7 +44,13 @@ Workflow
 
 2. Initialize a model with::
     
-      model = mb.Core(param = param_temp, max_space = 1000, data_name = 'data_in', alt=3)
+      model = mb.Core(
+          param = param_temp, 
+          data_name="artificial_data", 
+          alt=3,
+          equal_alt=1,
+          include_weights=False
+          )
       
    The structure of the input data and the parameter-input are given below.
 
@@ -62,43 +64,26 @@ during instantiation and within the estimation-method itself.
 | **Arguments for instantiation** (ov.Core(...)):
 | dict param: Indicated the names of the model attributes. The attribute-names 
 |       shall be derived from the column names of the input data.
-| int max_space: Defines the maximum number of parameter points within the 
-|       considered parameter space.
 | str data_name: Indicates the name of the input data-file. 
 | int alt: Indicates the number of considered choice alternatives.
 | int equal_alt: Indicates the maximum number of equal choice alternatives per choice set.
 |
 | **Keyword-arguments for instantiation** (ov.Core(...)): 
-| int sample_data: Define subset of data for estimation of base MNL-model.
-| tuple select_data: Defines a tuple of (str: attribute name, int/float: attribute value)
-|     to estimate mixed logit model upon specific subset of data.
 | boolean include_weights: If this is set to True, the model will search for a
 |     column in the input-data, called "weight", which indicates the weight
 |     for each observation. Defaults to True.
 | str initial_point_name: Specify name of pickle-file within subfolder *ModelParam*,
 |     if MNL-model was previously estimated (saves CPU-time).
-| dict param_transform: Specify alternative definition of parameter structure,
-|     if a previously estimated initial point (MNL-model) is loaded, but 
-|     further estimation is carried out upon alternative param-structure. 
-|     (e.g.: Parameters are set to random, which were fixed before.)
 |
 | **Keyword-arguments for estimation-method (model.estimate_mixed_logit(...))**:
 | int min_inter: Min. iterations for EM-algorithm.
-| int max_iter: Max. iterations for EM-algorithm
-| bool gpu: If set to True, then the estimation process is conducted using GPU-hardware.
-|     A necessary pre-condition for that is the existence of some NVIDIA-GPU
-|     on your hardware ans the installation of the python module numba.
+| int max_iter: Max. iterations for EM-algorithm.
+| float tol: Numerical tolerance of EM-algorithm.
 | bool bit_64: Defaults to False. If set to True, all numbers are calculated
 |     in 64-bit format, which increases precision, but also runtime.
-| str space_method: 'abs_value' or 'std_value', defines type of space-creation
-| int blind_exploration: Number of iteration in which shares are not updated. 
-|     Reduces initialization bias of random draws.
[…]
