+++ MANIFEST.in
@@ -1,4 +1,0 @@
-# -*- coding: utf-8 -*-
-
-graft /open_vector/InputData 
-graft /open_vector/ModelParam 
