### 2023-01-26 73ac61d
Updated README to include the option to install mode-behave via pip
