### 2023-01-30 d168439
Adapted post-processing routines to work with streamlit application
