+++ reference_environment.yml
@@ -8,6 +8,7 @@ dependencies:
   - pandas>=1.4.2
   - pickleshare>=0.7.5
   - scikit-learn>=1.0.2
+  - scikit-optimize>=0.9.0
   - scipy>=1.7.3
   - seaborn>=0.11.2
   
