### 2023-02-03 e17faa6
Initialization of k-means with reproducible results
