+++ mode_behave_public/simulation.py
@@ -173,7 +173,7 @@ class Simulation:
             prob_temp = calculate_logit_fast_MNL_cars(
                 i, 
                 hh_data, 
-                self.initial_point_cars, 
+                self.initial_point, 
                 asc_offset)
             probs.append(prob_temp)
         
@@ -470,7 +470,386 @@ class Simulation:
         
         probs = []
         for i in all_alternatives:
-            prob_temp = calculate_logit_fast_MNL_mode(i, trip_data, self.initial_point_mode, av)
+            prob_temp = calculate_logit_fast_MNL_mode(i, trip_data, self.initial_point, av)
             probs.append(prob_temp)
         
         return probs
+    
+    def simulate_hh_car_type_germany(self,
+                    AV,
+                    SEGMENT,
+                    RELATIVE_B_KOSTEN_MAL100,
+                    REICHWEITE_DURCH100,
+                    LADE_TANK_ZEIT,
+                    DISTANZ_LADE_TANK,
+                    CO2_MAL10,
+                    MEAN_HH_AGE,
+                    HH_SIZE,
+                    POPULATION,
+                    CAR_PARK_EV,
+                    CARSHARING,
+                    LONG_RANGE_AV,
+                    EV_EXPERIENCE,
+                    OWN_POWER,
+                    HH_OCC,
+                    PT_QUALITY,
+                    RELATIVER_KAUFPREIS,
+                         **kwargs):
+        """
+        This method calculates the probability of a car to have a certain 
+        propulsion type on hh-level.
+        0: 'ICEV'
+        1: 'PHEV'
+        2: 'BEV'
+        3: 'FCEV'
+            
+        Parameters
+        ----------
+        #   SEGMENT 0-3: Farhzeug-Segment nach Antrieb (Sollte gleich sein)
+        #   RELATIVE_B_KOSTEN_MAL100_0-3: Betriebskosten geteilt durch HH-Einkommen mal 100, nach Antrieb
+        #   REICHWEITE_DURCH100_0-3: Reichweite nach Antrieb, geteilt durch 100
+        #   LADE_TANK_ZEIT_0: Lade- bzw. Tankzeit nach Antrieb (N.A. für ICEV & PHEV)
+        #   DISTANZ_LADE_TANK_0: Distanz zur nächsten Ladesäule, Tankstelle nach Antrieb. (N.A. für ICEV & PHEV)
+        #   CO2_MAL10_0-3: Eingesparte CO2-Emissionen nach Antrieb.
+        #   MEAN_HH_AGE_0-3: Durchschnittsalter des Haushaltes
+        #   HH_SIZE_0-3: Anzahl der Haushaltsmitglieder
+        #   POPULATION_0-3: Einwohnerzahl des Wohnortes (Kreisebene) in 7 Kategorien:
+        #       (in Tausend) - 1: >500, 2:100-500 3:50-100, 4:20-50, 5:10-20, 6:5-10, 7:<5
+        #   CAR_PARK_EV_0-3: Ist es möglich, ein EV privat zu parken und zu laden? (I/O)
+        #   CARSHARING_0-3: Gibt es eine Carsharing-Mitgliedschaft im Haushalt? (I/O)
+        #   LONG_RANGE_AV_0-3: Gibt es ein weiteres Auto im HH mit "hoher Reichweite" (I/O)
+        #   EV_EXPERIENCE_0-3: Gibt es bereits ein EV im HH? (I/O)
+        #   OWN_POWER_0-3: Wird Strom teilweise selbst produziert, mit Wind. oder PV-Anlage? (I/O)
+        #   HH_OCC_0-3: Wie viele Voll- und Teilzeitarbeitnehmer gibt es im HH?
[…]
