### 2023-02-03 bc2dcf2
Updated manifest.in and readme.rst
