+++ mode_behave_public/post_analysis.py
@@ -902,7 +902,286 @@ class PostAnalysis:
         else:
             raise ValueError('No such method defined.')
             
-           
+    def visualize_all_attributes(self, **kwargs):
+        """
+        This method visualizes all attribute weights and additionally 
+        indicates the t-statistics, based on the
+        estimation results of the standard logit model.
+
+        Parameters
+        ----------
+        
+        kwargs save_fig_path : string
+            Path, which indicated the place where to store the visualization
+            as a .png-file.
+            
+        kwargs names_choice_options : dict
+            If given, this shall be a dictionary, which holds the 
+            names of the choice options as values and the numerical
+            indication of the choice option (0,1,2,...) as keys.
+
+        kwargs shift_t_stats : float
+            Shifts the text-fields which indicate the value of the t-statistic
+            in positive or negative direction..
+
+        Returns
+        -------
+        None.
+
+        """
+        #get keyword arguments
+        model_name = kwargs.get('model_name', '')
+        save_fig_path = kwargs.get('save_fig_path', False)        
+        names_choice_options = kwargs.get('names_choice_options', False)  
+        control_left_annotation = kwargs.get('control_left_annotation', 0)  
+        control_right_annotation = kwargs.get('control_right_annotation', 0)  
+        control_legend = kwargs.get("control_legend", [0,0])  
+        ext_x_limits = kwargs.get("ext_x_limits", False)
+        control_colorbar_label = kwargs.get("control_colorbar_label", [0,0])
+
+        relative = kwargs.get('relative', False)
+        stepwise = kwargs.get('stepwise', False)
+        set_title = kwargs.get('set_title', False)
+        set_xlabel = kwargs.get('set_xlabel', False)
+        map_y_values = kwargs.get('map_y_values', False)
+        if names_choice_options == False:
+            names_choice_options = ["Choice option " + str(i) for i in range(self.count_c)]
+        
+        #get data
+        y_values = []
+        for attr in self.param['constant']['fixed']:
+            y_values.append(attr)
+        for attr in self.param['constant']['random']:
+            y_values.append(attr)
+        for attr in self.param['variable']['fixed']:
+            y_values.append(attr)
+        for attr in self.param['variable']['random']:
+            y_values.append(attr)
+            
+        x_values = {c:[] for c in range(self.count_c)}
+        x_values_mean = {c:[] for c in range(self.count_c)}
+        t_stats = {c:[] for c in range(self.count_c)}
+        for attr in y_values:
+            index_attribute = self.get_index_of_attribute(attr)
+            list_value_temp = []
+            for c in range(self.count_c):
+                list_value_temp.append(self.initial_point[index_attribute[c]])
+            max_list_value_temp = abs(max(list_value_temp, key=abs))
+            for c in range(self.count_c):
+                if relative:
+                    data_mean = self.data[attr+"_"+str(c)+"_0"].mean()
+                    utility_temp = list_value_temp[c]*data_mean
+                    x_values[c].append(utility_temp)
+                elif stepwise:
+                    #calculate mean value
+                    data_mean = self.data[attr+"_"+str(c)+"_0"].mean()
+                    utility_temp = list_value_temp[c]*data_mean
+                    x_values_mean[c].append(utility_temp)
+                    
+                    #calculate list of utilities
+                    unique_values = self.data[attr+"_"+str(c)+"_0"].unique()
+                    if len(unique_values) <= 10:
+                        unique_values.sort()
+                    else:
+                        val_max = max(unique_values)
+                        val_min = min(unique_values)
+                        step_temp = (val_max - val_min) / 10
+                        unique_values = np.arange(
[…]
