### 2023-02-07 ddd7d21
Updated core.py to enable exogenous definition of initial points
