+++ reference_environment.yml
@@ -2,6 +2,8 @@ name: env_mode_behave
 channels:
   - conda-forge
 dependencies:
+  - python=3.8
+  - Pillow>=9.4.0
   - matplotlib>=3.5.1
   - numba>=0.55.1
   - numpy>=1.21.5
@@ -11,5 +13,3 @@ dependencies:
   - scikit-optimize>=0.9.0
   - scipy>=1.7.3
   - seaborn>=0.11.2
-  
-prefix: C:\Users\j.reul\.conda\envs\env_mode_behave
