### 2023-03-16 #5265 @editorialbot
[REVIEW]: MODE.behave: A Python Package for Discrete Choice Modeling
**Submitting author:** @julianreul (<a href="http://orcid.org/0000-0002-2786-0233">Julian Reul</a>)
**Repository:** https://github.com/julianreul/mode_behave
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.13
