### 2023-03-17 #5265 @l-kotzur
@julianreul

>  it is the first to provide functionality for the estimation of mixed logit models with nonparametric distribution

What is the advantage or benefit of a nonparametric distribution to a parametric distributions of the other packages for the application cases?
