### 2023-03-17 #5265 @l-kotzur
@julianreul Could you add the code on how to reproduce the plots in the imgs-folder?

Can you use English names instead of German names for the parameters in example_estimation.py?

Can you either add the data and for the Japanese market or drop the plots? They are not reproducible.

Can you remove dist folder from the repo? It makes the package unnecessarily large.
