### 2023-03-17 #5265 @l-kotzur
@juliuanreul Could you modify the core module (or extract)  such that the input data and model parameters can be directly provided either as numpy arrays or pandas dataframes? In its current application it cumbersome to integrate the package in an workflow since that data cannot be directly be given to the model.

Can you explain the .picke files in ModelParam? Are they necessary for the flow? If so, could you use a human-readable datatype such as .json or .csv? In their current form it is not transparent how they integrate to the model.
