### 2023-03-21 50e466f
Modified paper.md according to comments by L. Kotzur (I)
