### 2023-03-21 0a869e3
Deleted imgs-folder and shifted data to mode_behave_public/Visualizations (answer to eview comment L. Kotzur)
