### 2023-03-21 603119e
Changed example_estimation.py to include English model parameters and to pass the input data directly to the created model as an argument.
