### 2023-03-21 f6ae93e
Deleted unnecessary files (dist-folder, artificial_data.csv)
