### 2023-03-21 fd6813b
Substituted all .pickle-files with .csv-files and adapted the code accordingly
