+++ test/test_estimation.py
@@ -52,7 +52,7 @@ class TestEstimation(unittest.TestCase):
             )
         dist_c = stats.multivariate_normal(mean=mu_c, cov=cov_c)
         #____obtain random sample from copula distribution
-        sample_c = dist_c.rvs(size=size_dataset)
+        sample_c = dist_c.rvs(size=size_dataset, random_state=42)
         #____obtain marginals from copula distribution
         choice = sample_c[:,0]
         attr_x = sample_c[:,1]
@@ -166,10 +166,12 @@ class TestEstimation(unittest.TestCase):
         param_temp['variable']['fixed'] = param_fixed
         param_temp['variable']['random'] = param_random   
             
+        artificial_data = self.get_artificial_data()
+        
         #Initialize model
         model = mb.Core(
             param=param_temp, 
-            data_name="artificial_data", 
+            data_in=artificial_data, 
             alt=3,
             equal_alt=1,
             include_weights=False,
@@ -187,15 +189,10 @@ class TestEstimation(unittest.TestCase):
             t_stats_out=False
             )
         
-        with open(model.PATH_ModelParam + "initial_point_artificial_data.pickle", 'rb') as handle:
-            initial_point_compare = pickle.load(handle)  
-                
-        with open(model.PATH_ModelParam + "shares_artificial_data.pickle", 'rb') as handle:
-            shares_compare = pickle.load(handle)  
-            
-        with open(model.PATH_ModelParam + "points_artificial_data.pickle", 'rb') as handle:
-            points_compare = pickle.load(handle)  
-        
+        initial_point_compare = np.genfromtxt(model.PATH_ModelParam + "initial_point_artificial_data.csv", delimiter=",")
+        shares_compare = np.genfromtxt(model.PATH_ModelParam + "shares_artificial_data.csv", delimiter=",")
+        points_compare = np.genfromtxt(model.PATH_ModelParam + "points_artificial_data.csv", delimiter=",")
+
         #test estimation of initial_point (via the method estimate_logit())
         self.assertTrue(np.allclose(model.initial_point, initial_point_compare, atol=0.1))
         
@@ -203,7 +200,7 @@ class TestEstimation(unittest.TestCase):
         self.assertTrue(np.allclose(model.shares, shares_compare, atol=0.1))
         
         #test definition of parameter space (points)
-        self.assertTrue(np.allclose(model.points, points_compare, atol=0.1))
+        self.assertTrue(np.allclose(np.array(model.points), points_compare, atol=0.1))
 
         
     def test_estimate_logit(self):
@@ -234,10 +231,12 @@ class TestEstimation(unittest.TestCase):
         param_temp['variable']['fixed'] = param_fixed
         param_temp['variable']['random'] = param_random         
             
+        artificial_data = self.get_artificial_data()
+        
         #Initialize model
         model = mb.Core(
             param=param_temp, 
-            data_name="artificial_data", 
+            data_in=artificial_data, 
             alt=3,
             equal_alt=1,
             include_weights=False,
@@ -245,8 +244,7 @@ class TestEstimation(unittest.TestCase):
                  
         model.initial_point = model.estimate_logit(stats=False)
         
-        with open(model.PATH_ModelParam + "initial_point_artificial_data.pickle", 'rb') as handle:
-            initial_point_compare = pickle.load(handle)  
+        initial_point_compare = np.genfromtxt(model.PATH_ModelParam + "initial_point_artificial_data.csv", delimiter=",")
                 
         #test estimation of initial_point (via the method estimate_logit())
         self.assertTrue(np.allclose(model.initial_point, initial_point_compare, atol=0.1))
