+++ .github/workflows/testing.yml
@@ -1,5 +1,3 @@
-name: Python package
-
 on: [push]
 
 jobs:
@@ -16,12 +14,8 @@ jobs:
         uses: actions/setup-python@v4
         with:
           python-version: ${{ matrix.python-version }}
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install unittest
-          if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
       - name: Test with unittest
         run: |
           python -m unittest test_estimation.py
           python -m unittest test_estimation.py
+          
