+++ .github/workflows/testing.yml
@@ -19,6 +19,7 @@ jobs:
           if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
       - name: Test with unittest
         run: |
+          pip install -e .
           cd test
           python -m unittest test_estimation.py
           python -m unittest test_estimation.py
