### 2023-03-21 #5265 @julianreul
Dear @l-kotzur , thanks for your feedback. - I managed to work on most of your comments. Changes are commited to the "review"-branch of the repo.

> Shall https://github.com/julianreul/mode_behave or https://github.com/FZJ-IEK3-VSA/mode_behave be the place of the repository? (DONE)

The place of the reference-repository shall be https://github.com/julianreul/mode_behave, as this was indicated as the reference-repo during the submission process.

> What is the advantage or benefit of a nonparametric distribution to a parametric distributions of the other packages for the application cases? (DONE)

Parametric mixed logit models exogenously assume the shape of the preference distribution, which is defined to be normal or log-normal in most cases. Thus, the analysis of preference heterogeneity is limited to that pre-defined shape. Nonparametric mixed logit models are unbiased of such assumptions and derive the (discrete) preference distribution in the base population endogenously from the data. This facilitates the analysis of more complex (discrete) distributions and, based on that, the segregation of preference groups with diverging characteristics.
I added an explanation of the advantages and disadvantages in the paper.md.

> Can you use English names instead of German names for the parameters in example_estimation.py? (DONE)

Yes, I changed the parameter names to English.

> Can you either add the data for the Japanese market or drop the plots? They are not reproducible. (DONE)

I dropped the plots.

> Can you remove dist folder from the repo? It makes the package unnecessarily large. (DONE)

I removed the dist folder.

> Can you integrate automated runs of your tests scripts, e.g., to Github workflow? (DONE)

I integrated automated tests via GitHub actions. The respective GitHub action is working and executes the tests defined in the test-folder.

> Could you modify the core module (or extract) such that the input data and model parameters can be directly provided either as numpy arrays or pandas dataframes? In its current application it cumbersome to integrate the package in an workflow since that data cannot be directly be given to the model.

I changed the functionality in the core module. Input data can now be provided via the keyword-argument "data_in" as pandas dataframes, which are loaded from .csv-files (or any other data format). The example_estimation.py script and the readme.txt have been adapted accordingly.

> Can you explain the .picke files in ModelParam? Are they necessary for the flow? If so, could you use a human-readable datatype such as .json or .csv? In their current form it is not transparent how they integrate to the model.

The files in the folder ModelParam contain pre-estimated model parameters, which are required for the functionality of the simulation-routines (inherently based on pre-estimated logit models) and the estimation-routines.
However, I substituted all .pickle-files by .csv files and adapted the code accordingly.

> Could you add contribution guidelines to the repo? (OPEN)

This point is still open, I will work on it towards the end of the week.

Thanks again for your comments.
