+++ contributing.md
@@ -1,0 +1,88 @@
+# Contributing to MO|DE.behave
+
+## Welcome
+
+Thanks for taking time to read this contribution-file and for your 
+intent to contribute to the project!
+
+## Contribution Process
+
+* Before making a contribution, please contact the project owner via issue or email. 
+We aim for a short exchange on your contribution-ideas to streamline the subsequent 
+contribution and to check for coherency with the project's targets (next item).
+* Please continue with a usual "fork and pull request" workflow to implement
+your contribution. If you contacted the project maintainer via issue, 
+please reference the issue when submitting your pull request.
+* If you add new functionality to the project, please make sure to cover
+your contribution with appropriate tests. Furthermore, please ensure,
+that the existing tests remain successful.
+* Changes in the functionality of the model should be stated/adapted
+in the readme-file.
+
+## Project Targets
+
+* Central targets: 
+    - Provision of an open-source solution for
+the rapid implementation and estimation of discrete choice models.
+    - Promotion of the mixed logit model with nonparametric design.
+    - Integration of additional visualization and simulation methods.
+* Future development targets:
+    - Integration of mixed logit models with parametric design.
+    - Automatic compatability with standard DCE output files (e.g., from sawtooth-software)
+* Long-term vision: Becoming the **building block of agent-based frameworks**.
+
+## Our Pledge
+
+In the interest of fostering an open and welcoming environment, we as
+contributors and maintainers pledge to make participation in our project and
+our community a harassment-free experience for everyone, regardless of age, body
+size, disability, ethnicity, sex characteristics, gender identity and expression,
+level of experience, education, socio-economic status, nationality, personal
+appearance, race, religion, or sexual identity and orientation.
+
+## Our Standards
+
+Examples of behavior that contributes to creating a positive environment
+include:
+
+* Using welcoming and inclusive language
+* Being respectful of differing viewpoints and experiences
+* Gracefully accepting constructive criticism
+* Focusing on what is best for the community
+* Showing empathy towards other community members
+
+## Our Responsibilities
+
+Project maintainers are responsible for clarifying the standards of acceptable
+behavior and are expected to take appropriate and fair corrective action in
+response to any instances of unacceptable behavior.
+
+Project maintainers have the right and responsibility to remove, edit, or
+reject comments, commits, code, wiki edits, issues, and other contributions
+that are not aligned to this Code of Conduct, or to ban temporarily or
+permanently any contributor for other behaviors that they deem inappropriate,
+threatening, offensive, or harmful.
+
+## Enforcement
+
+Instances of abusive, harassing, or otherwise unacceptable behavior may be
+reported by contacting the project team at julian.reul@rwth-aachen.de. All
+complaints will be reviewed and investigated and will result in a response that
+is deemed necessary and appropriate to the circumstances. The project team is
+obligated to maintain confidentiality with regard to the reporter of an incident.
+Further details of specific enforcement policies may be posted separately.
+
+Project maintainers who do not follow or enforce the Code of Conduct in good
+faith may face temporary or permanent repercussions as determined by other
+members of the project's leadership.
+
+## Attribution
+
+This Code of Conduct is adapted from the [Contributor Covenant][homepage], version 1.4,
+available at https://www.contributor-covenant.org/version/1/4/code-of-conduct.html
+
+[homepage]: https://www.contributor-covenant.org
+
+For answers to common questions about this code of conduct, see
+https://www.contributor-covenant.org/faq
+
