+++ .github/workflows/testing.yml
@@ -16,9 +16,10 @@ jobs:
           python-version: ${{ matrix.python-version }}
       - name: Install dependencies
         run: |
-          if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+          conda env update -q --file=reference_environment.yml
       - name: Test with unittest
         run: |
+          conda activate env_mode_behave
           pip install -e .
           cd test
           python -m unittest test_estimation.py
