+++ .github/workflows/testing.yml
@@ -16,10 +16,9 @@ jobs:
           python-version: ${{ matrix.python-version }}
       - name: Install dependencies
         run: |
-          conda env update -q --file=reference_environment.yml
+          pip install -r requirements.txt
       - name: Test with unittest
         run: |
-          conda activate env_mode_behave
           pip install -e .
           cd test
           python -m unittest test_estimation.py
