+++ .github/workflows/testing.yml
@@ -22,5 +22,5 @@ jobs:
           pip install -e .
           cd test
           python -m unittest test_estimation.py
-          python -m unittest test_estimation.py
+          python -m unittest test_simulation.py
           
