+++ reference_environment.yml
@@ -11,5 +11,5 @@ dependencies:
   - pickleshare>=0.7.5
   - scikit-learn>=1.0.2
   - scikit-optimize>=0.9.0
-  - scipy>=1.7.3
+  - scipy=1.9.3
   - seaborn>=0.11.2
