+++ .github/workflows/testing.yml
@@ -17,15 +17,14 @@ jobs:
       - name: Install dependencies
         run: |
           pip install -r requirements.txt
-      - name: Test simulation
-        run: |
           pip install -e .
           cd test
+
+      - name: Test simulation
+        run: |
           python -m unittest test_simulation.py
       - name: Test estimation
         run: |
-          pip install -e .
-          cd test
           python -m unittest test_estimation.py
 
           
