+++ mode_behave_public/estimation.py
@@ -708,7 +708,7 @@ class Estimation:
             loglike, 
             x0, 
             method="L-BFGS-B", 
-            tol=1e-6, 
+            tol=1e-9, 
             jac='cs'
             )
         end_logit = time.time()
@@ -743,7 +743,7 @@ class Estimation:
                     loglike, 
                     x0, 
                     method="L-BFGS-B", 
-                    tol=1e-6, 
+                    tol=1e-9, 
                     jac='cs')
                 #iterate over estimated coefficients
                 for j, param in enumerate(res.x):
