### 2023-03-26 #5265 @julianreul
> Could you add contribution guidelines to the repo? (DONE)

I added a contributing.md file to the repo, according to my ideas of how to contribute.
