### 2023-03-27 #5265 @julianreul
@l-kotzur Thanks for your feedback, it definitely helped to improve the submission!

> I thought that my new role would not be seen as COI but the rules are quite clear on this situation "In addition, your recent (past year) association with the same organization of a submitter is a COI, for example, being employed at the same institution."

I should have checked these regulations more thoroughly before.

I hope we can find another reviewer soon.
