### 2023-04-11 #5265 @julianreul
Dear @sbenthall, I contacted a few potential reviewers, but received negative feedback.

Another potential reviewer would be Joe Molloy (Github-name: joemolloy), who developed the mixl-package at the ETH Zürich. The mixl-package is a framework for estimating discrete choice models, similar to MO|DE.behave, but written in R.
Thus, it would be a good fit regarding the theoretical background.
