### 2023-04-13 #5265 @sbenthall
Hello @joemolloy,

Would you be willing to review this submission to the Journal of Open Source Software?

Its author says you have written a similar library in R to this submission.

We are looking for a second reviewer after one of the reviewers backed out because of a conflict of interest.

Thanks,
@sbenthall
