+++ .github/workflows/testing.yml
@@ -1,0 +1,28 @@
+on: [push]
+
+jobs:
+  build:
+
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ["3.8"]
+
+    steps:
+      - uses: actions/checkout@v3
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v4
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Install dependencies
+        run: |
+          pip install -r requirements.txt
+          pip install -e .
+      - name: Test simulation
+        run: |
+          cd test
+          python -m unittest test_simulation.py
+      - name: Test estimation
+        run: |
+          cd test
+          python -m unittest test_estimation.py
