+++ test/test_simulation.py
@@ -11,6 +11,7 @@ import numpy as np
 import mode_behave_public as mb
 
 class TestCarOwnership(unittest.TestCase):
+    
     def test_simulate_hh_cars(self):
         """
         Integration test of the method -simulate_hh_cars()-
@@ -30,7 +31,7 @@ class TestCarOwnership(unittest.TestCase):
         regiontype = 1
         
         #simulate
-        model = mb.Core(model_type = 'simulation')        
+        model = mb.Core(model_type = 'simulation', simulation_type = 'car_ownership')        
         result = model.simulate_hh_cars(urban_region, rural_region, hh_size,
                              adults_working, children, htype, quali_opnv, sharing,
                              relative_cost_per_car, age_adults, regiontype)
@@ -46,13 +47,13 @@ class TestCarOwnership(unittest.TestCase):
             atol=0.01)
             )      
         
-    def test_simulate_mode_choice(self): 
+    def test_simulate_mode_choice(self):
         """
         Integration test of the method -test_simulate_mode_choice()-.
 
         """
         #simulate
-        model = mb.Core(model_type = 'simulation')        
+        model = mb.Core(model_type = 'simulation', simulation_type = 'mode_choice')        
         result = model.simulate_mode_choice(
             agegroup=2,
             occupation=1,
