+++ test/test_estimation.py
@@ -136,6 +136,7 @@ class TestEstimation(unittest.TestCase):
             
         return dataset
            
+    
     def test_estimate_logit(self):
         """
         Integration test of the method -estimate_logit()-
@@ -186,73 +187,70 @@ class TestEstimation(unittest.TestCase):
         self.assertTrue(np.allclose(model.initial_point, initial_point_compare, atol=0.1))
     
     
-    
-# =============================================================================
-#     def test_estimate_mixed_logit(self):
-#         """
-#         Integration test of the method -estimate_mixed_logit()-
-#         Utilized CPU hardware only.
-# 
-#         """
-#         param_fixed = []
-#         param_random = [
-#             "attr_x",
-#             "attr_y",
-#             "attr_z"
-#             ]
-#         
-#         param_temp = {'constant': 
-#                           {
-#                            'fixed':[],
-#                            'random':[]
-#                            },
-#                       'variable':
-#                           {
-#                            'fixed': [],
-#                            'random':[]
-#                            }
-#                       }
-#         
-#         param_temp['variable']['fixed'] = param_fixed
-#         param_temp['variable']['random'] = param_random   
-#             
-#         artificial_data = self.get_artificial_data()
-#         
-#         #Initialize model
-#         model = mb.Core(
-#             param=param_temp, 
-#             data_in=artificial_data, 
-#             alt=3,
-#             equal_alt=1,
-#             include_weights=False,
-#             )
-#         
-#         #estimate mixed logit model
-#         model.estimate_mixed_logit(
-#             min_iter=10, 
-#             max_iter=1000,
-#             tol=0.01,
-#             space_method = 'std_value',
-#             scale_space = 2,
-#             max_shares = 1000,
-#             bits_64=True,
-#             t_stats_out=False
-#             )
-#         
-#         initial_point_compare = np.genfromtxt(model.PATH_ModelParam + "initial_point_artificial_data.csv", delimiter=",")
-#         shares_compare = np.genfromtxt(model.PATH_ModelParam + "shares_artificial_data.csv", delimiter=",")
-#         points_compare = np.genfromtxt(model.PATH_ModelParam + "points_artificial_data.csv", delimiter=",")
-# 
-#         #test estimation of initial_point (via the method estimate_logit())
-#         self.assertTrue(np.allclose(model.initial_point, initial_point_compare, atol=0.1))
-#         
-#         #test estimation of shares
-#         self.assertTrue(np.allclose(model.shares, shares_compare, atol=0.1))
-#         
-#         #test definition of parameter space (points)
-#         self.assertTrue(np.allclose(np.array(model.points), points_compare, atol=0.1))
-# 
-# =============================================================================
+    def test_estimate_mixed_logit(self):
+        """
+        Integration test of the method -estimate_mixed_logit()-
+        Utilized CPU hardware only.
+
+        """
+        param_fixed = []
+        param_random = [
+            "attr_x",
+            "attr_y",
+            "attr_z"
+            ]
+        
+        param_temp = {'constant': 
+                          {
+                           'fixed':[],
+                           'random':[]
+                           },
+                      'variable':
+                          {
+                           'fixed': [],
+                           'random':[]
+                           }
+                      }
+        
+        param_temp['variable']['fixed'] = param_fixed
+        param_temp['variable']['random'] = param_random   
+            
+        artificial_data = self.get_artificial_data()
+        
+        #Initialize model
+        model = mb.Core(
+            param=param_temp, 
+            data_in=artificial_data, 
+            alt=3,
+            equal_alt=1,
+            include_weights=False,
[…]
