### 2023-05-02 #5265 @julianreul
Hello @sbenthall,

today I received feedback from another researcher at the DLR (German Aerospace Center), who is willing to review our submission. Her name is Fabia Miorelli (GitHub-username: **fabmio**).

She is a PhD student at the DLR, working intensively with Python in the field of energy systems analysis and co-published another open-source model in Python, called "vencopy" (Gitlab: https://gitlab.com/dlr-ve/esy/vencopy/vencopy).

I would appreciate, if you nominate her as a reviewer to progress with the review-status.

Thanks,
@julianreul
