### 2023-07-02 julianreul/mode_behave#5 @abhiramm7
JOSS Review: Documentation
Hi,

The package looks fantastic! Awesome work 🙌🏽.

Just a couple of minor comments,

1. The manuscript indicates that this package can leverage GPUs, I might have missed something, but there is no reference to how the users can use this functionality. If the package is built to use the machine's GPU automatically, please indicate it in the documentation.
2. Testing: Please add the testing steps to the readme file in correspond with JOSS guidelines.
3. Update `Arguments for instantiation` , `Keyword-arguments for instantiation`, and `Keyword-arguments for estimation-method` documentation with code formatting. That would make it easier to follow. Same for `Structure of Parameters and Input Data`.

Thanks,
Abhi
