### 2023-07-02 julianreul/mode_behave#6 @abhiramm7
JOSS Review: Example Script
Hi,

Consider updating the `example_estimation.py` script with visualization examples. I might have missed something, but I couldn't find an example of how to use the visualization functionality in the package. I would also recommend adding the script used to generate the figures presented in the manuscript.

Thanks,
Abhi
