### 2023-07-02 #5265 @abhiramm7
Hi @julianreul , I've created two issues https://github.com/julianreul/mode_behave/issues/6 and https://github.com/julianreul/mode_behave/issues/5 in the project repository. Overall the package looks amazing. I just had a couple of minor comments, I hope you find them useful. Thanks! :D.
