+++ README.rst
@@ -76,28 +76,31 @@ Workflow
 The estimation of the mixed logit model can be modified by definition of keyword-arguments
 during instantiation and within the estimation-method itself.
 
-| **Arguments for instantiation** (ov.Core(...)):
-| dict param: Indicated the names of the model attributes. The attribute-names 
-|       shall be derived from the column names of the input data.
-| str data_name: Indicates the name of the input data-file. 
-| int alt: Indicates the number of considered choice alternatives.
-| int equal_alt: Indicates the maximum number of equal choice alternatives per choice set.
-|
-| **Keyword-arguments for instantiation** (ov.Core(...)): 
-| boolean include_weights: If this is set to True, the model will search for a
-|     column in the input-data, called "weight", which indicates the weight
-|     for each observation. Defaults to True.
-|
-| **Keyword-arguments for estimation-method (model.estimate_mixed_logit(...))**:
-| int min_inter: Min. iterations for EM-algorithm.
-| int max_iter: Max. iterations for EM-algorithm.
-| float tol: Numerical tolerance of EM-algorithm.
-| bool bit_64: Defaults to False. If set to True, all numbers are calculated
-|     in 64-bit format, which increases precision, but also runtime.
-| str space_method: Defines the chosen method to span the parameter space for the mixed logit estimation.
-| int scale_space: Defines the size of the space, relative to the chosen space_method.
-| int max_shares: Defines the maximum number of points to be observed in the parameter space.
-|
+**Arguments for instantiation** (ov.Core(...))::
+
+    dict param: Indicated the names of the model attributes. The attribute-names 
+      shall be derived from the column names of the input data.
+    str data_name: Indicates the name of the input data-file. 
+    int alt: Indicates the number of considered choice alternatives.
+    int equal_alt: Indicates the maximum number of equal choice alternatives per choice set.
+
+**Keyword-arguments for instantiation** (ov.Core(...))::
+
+    boolean include_weights: If this is set to True, the model will search for a
+        column in the input-data, called "weight", which indicates the weight
+        for each observation. Defaults to True.
+
+**Keyword-arguments for estimation-method (model.estimate_mixed_logit(...))**::
+
+    int min_inter: Min. iterations for EM-algorithm.
+    int max_iter: Max. iterations for EM-algorithm.
+    float tol: Numerical tolerance of EM-algorithm.
+    bool bit_64: Defaults to False. If set to True, all numbers are calculated
+        in 64-bit format, which increases precision, but also runtime.
+    str space_method: Defines the chosen method to span the parameter space for the mixed logit estimation.
+    int scale_space: Defines the size of the space, relative to the chosen space_method.
+    int max_shares: Defines the maximum number of points to be observed in the parameter space.
+
       
 4. The package can also be used to estimate multinomial logit models::
 
@@ -111,6 +114,42 @@ during instantiation and within the estimation-method itself.
 
     PATH_TO_PACKAGE/mode_behave_public/Deployments/example_estimation.py
 
+Testing
+=======
+
+The software includes testing routines, written with the package -unittest-, 
+to ensure its functionality throughout the development process. 
+The first test-routine checks the functionality
+of the estimation routines (PATH: ./test/test_estimation.py), while the second
+test routine checks the functionality of simulation routines 
+(PATH: ./test/test_simulation.py)
+
+These testing routines can be activated in two ways:
+
+1. Via GitHub Actions:
+    Whenever a new commit is pushed to the repository, GitHub Actions
+    are automatically triggered, which execute the test routines.
+    The test results are displayed in the GitHub Actions tab in the 
+    software's repository online.
+2. Via manual execution:
+    Alternatively, the test routines can be called manually. You might chose
+    this option, if you develop the software locally and want to validate 
[…]
