+++ README.rst
@@ -79,7 +79,7 @@ during instantiation and within the estimation-method itself.
 **Arguments for instantiation (ov.Core(...))**::
 
     dict param:
-        Indicated the names of the model attributes. 
+        Indicates the names of the model attributes. 
         The attribute-names shall be derived from the column names of the input data.
     str data_name: 
         Indicates the name of the input data-file. 
@@ -170,50 +170,63 @@ Structure of Parameters and Input Data
 
    The input dataset contains the observations with which the model is 
    calibrated. The input data is called with the specified string of the
-   keyword-argument *data_name*. The input data shall be placed in .csv- or 
-   .pickle-format within the subfolder *InputData* of the package *mode_behave*.
-   The data shall follow the structure below:
+   keyword-argument *data_in*. The input data must be loaded from .csv- or 
+   .pickle-format before model initialization.
+   The data shall follow the structure below::
    
-   * Rows: Observations.
-   * Columns:
-         - One column per parameter of the utility function AND per alternative AND per equal alternative.
-           Specified as: 'Attribute_name_' + str(no_alternative) + str(no_equal_alternative)
-         - One column for the choice-indication of EACH alternative AND per equal alternative.
-           Specified as: 'choice_' + str(no_alternative) + str(no_equal_alternative)
-         - One column per alternative AND per equal alternative, indicating the availability.
-           Specified as: 'av_' + str(no_alternative) + str(no_equal_alternative)
-         - If a parameter is constant across alternatives or equal alternatives, then let the columns be equal.
-         - Furthermore, the observations can be given a weight. Therefore, an additional column needs to be provided, named 'weight'. - Without any further suffix.
-   * Index: The index shall start from '0'.
+       Rows: 
+           Observations.
+       
+       Columns:
+           One column per parameter of the utility function AND per alternative AND per equal alternative.
+           Specified as: **'Attribute_name_' + str(no_alternative) + str(no_equal_alternative)**
+           
+           One column for the choice-indication of EACH alternative AND per equal alternative.
+           Specified as: **choice_' + str(no_alternative) + str(no_equal_alternative)**
+           
+           One column per alternative AND per equal alternative, indicating the availability.
+           Specified as: **'av_' + str(no_alternative) + str(no_equal_alternative)**
+           
+           If a parameter is constant across alternatives or equal alternatives, then let the columns be equal.
+           
+           Furthermore, the observations can be given a weight. Therefore, an additional column needs to be provided, named 'weight'. - Without any further suffix.
+       
+       Index: The index shall start from '0'.
           
 2. Initialization argument 'param':
     
    'param' is specified as a dictionary containing the attribute names of the 
-   utility function, sorted by type.
+   utility function, sorted by type::
    
-   * param['constant']['fixed']: Attributes, which are constant over choice 
-     options and fixed within the parameter space. 
-   * param['constant']['random']: Attributes, which are constant over choice 
-     options and randomly distributed over the parameter space. 
-   * param['variable']['fixed']: Attributes, which vary over choice 
-     options and are fixed within the parameter space. 
-   * param['variable']['random']: Attributes, which vary over choice 
-     options and are randomly distributed over the parameter space. 
+   param['constant']['fixed']: 
+       Attributes, which are constant over choice 
+       options and fixed within the parameter space. 
+   param['constant']['random']: 
+       Attributes, which are constant over choice 
+       options and randomly distributed over the parameter space. 
[…]
