### 2023-07-03 julianreul/mode_behave#5 @julianreul
Hi,

thanks again for your comments on the package! I modified the package, especially the readme and the paper, accordingly and pushed the changes in recent commits. Here is an overview of the changes as one-by-one responses to your comments:

1. GPU-leverage: I incorporated GPU-leverage in an earlier version of the package. Later on I found, that the additional overhead of GPU-utilization was too large to significantly improve the model's performance. So, I deleted any reference to GPU-utilization from the paper. However, I improved the model performance by utilizing latin hypercube sampling within the expectation maximization algorithm. I added a comment in the paper accordingly.
2. Documentation on testing: I added a section called "Testing" in the readme file below the section "Workflow". In this section, I explain which testing routines are implemented in the package and how to use them (a: automatically via GitHub actions, b: manually via the Anaconda prompt)
3. Reformatting sections of the readme file: I reformatted the indicated sections to code formatting for better readability. Especially, the reference to (keyword-)arguments of the explained methods were reformatted.

I hope to have sufficiently addressed your comments in this issue. Please let me know, if there are any open points to address.

Best
Julian
