### 2023-07-03 36ce117
Changed example_estimation.py according to reviewer's comments #6
