### 2023-07-05 73050a3
Simplification of simulation routine .simulate_hh_cars()
