### 2023-07-05 27082fe
Added functionality in forecast() to integrate exogenous definition of availabilities
