+++ mode_behave_public/post_analysis.py
@@ -1341,6 +1341,13 @@ class PostAnalysis:
             for each choice option.
         kwargs asc_offset : list
             offset values for alternative specific constants
+        kwargs av_external : numpy array
+            This array is used to exogenously define the availabilities
+            for each choice option during a simulation.
+            The array must have as much entries as choice options are 
+            being observed: len(av_external) = self.count_c
+            Define the availability to 1 (always available), 0 (never available)
+            or np.nan (availability according to base data).
             
         Returns
         -------
@@ -1357,6 +1364,25 @@ class PostAnalysis:
             "asc_offset", 
             np.array([0 for c in range(self.count_c)], dtype="float64")
             )
+        av_external = kwargs.get("av_external", False)
+        
+        #exogeneous definiton of choice availabilities
+        if av_external:
+            #check for the correct number of exogenously defined availabilities.
+            if len(av_external) == self.count_c:
+                #interate over availabilities
+                for av_ext_count, av_ext in enumerate(av_external):
+                    if np.isnan(av_ext):
+                        continue
+                    else:
+                        if av_ext in [0,1]:
+                            self.av[av_ext_count] = av_ext
+                        else:
+                            raise ValueError("External availability must be 0 or 1.")
+            else:
+                raise AttributeError("Number of defined availabilities does not match number of choice options.")
+        else:
+            self.av = self.av_backup.copy()
 
         no_constant_fixed = len(self.param['constant']['fixed'])
         no_constant_random = len(self.param['constant']['random'])
@@ -1459,6 +1485,25 @@ class PostAnalysis:
         observations within a given base-sample.
         Requires prior call of estimate_mixed_logit().
 
+        Parameters
+        ----------
+        latent_points : 2D numpy array.
+            The random points within each class.
+        latent_shares : 1D numpy array.
+            The share of each class.
+        kwargs sense : dictionary
+            The dictionary "sense" holds the attribute names for which sensitivities
+            shall be simulated as keys. The values are the arrays or lists
+            which indicate the relative change of the attribute value
+            for each choice option.
+        kwargs av_external : numpy array
+            This array is used to exogenously define the availabilities
+            for each choice option during a simulation.
+            The array must have as much entries as choice options are 
+            being observed: len(av_external) = self.count_c
+            Define the availability to 1 (always available), 0 (never available)
+            or np.nan (availability according to base data).
+
         Returns
         -------
         PandasSeries
@@ -1471,7 +1516,26 @@ class PostAnalysis:
         sense = kwargs.get("sense", {})
         vector_output_no_weights = kwargs.get("vector_output_no_weights", False)
         asc_offset = kwargs.get("asc_offset", np.array([0 for c in range(self.count_c)], dtype="float64"))
+        av_external = kwargs.get("av_external", False)
         
+        #exogeneous definiton of choice availabilities
+        if av_external:
+            #check for the correct number of exogenously defined availabilities.
+            if len(av_external) == self.count_c:
+                #interate over availabilities
+                for av_ext_count, av_ext in enumerate(av_external):
+                    if np.isnan(av_ext):
+                        continue
+                    else:
+                        if av_ext in [0,1]:
+                            self.av[av_ext_count] = av_ext
+                        else:
+                            raise ValueError("External availability must be 0 or 1.")
+            else:
+                raise AttributeError("Number of defined availabilities does not match number of choice options.")
+        else:
+            self.av = self.av_backup.copy()
+            
         if mixing_distribution == "discrete":       
             initial_point = self.initial_point
             no_constant_fixed = self.no_constant_fixed
@@ -1619,6 +1683,13 @@ class PostAnalysis:
             shall be simulated as keys. The values are the arrays or lists
[…]
