### 2023-07-09 #5265 @fabmio
Hi @julianreul, I have completed the first iteration of my review, below you can find some comments and ideas:

**General checks**
- Documentation: Overall I think it would be really nice and easier to navigate to have a separate software documentation page (e.g. readthedocs) than the readme alone. This way it would also be possible to look up through the docstrings all methods in each class and their respective arguments from the documentation instead of looking into the codebase.

**Functionality**
- Code: There are lots of trailing whitespaces and white lines containing white spaces in the code as well as missing white spaces after block comments, missing white spaces after commas in arguments and continuation lines which are in some cases over and underindented. I would suggest to install an autoformatter and riformat the entire codebase.
- Code: In the estimation file on line 114 avoid the usage of a bare Except and on line 386 rename the variable l which is ambiguous (is it an I or small L?)
- Testing: I run both tests from terminal: the test_simulation ran without errors. In the test_estimation one test failed with the following error message:  AssertionError: False is not true on line 252 in test_estimate_mixed_logit (self.assertTrue(np.allclose(np.array(model.points), points_compare, atol=0.1))

**Documentation**
- Example usage: I would rewrite the example usage in the documentation to reflect 1:1 the example_estimation.py file so that can work as a guideline and introduction to the tool for new users (explicitly adding for example the pandas import)
- Contribution file: In the contribution file I would add that the new contributions should also be properly documented.
- Visualisations: it would be nice to have also the visualisation module documented.
- Issue reporting: I would indicate in the documentation where users could report issues or problems with the software (issue board, email, online channel like gitter, slack, or similar? an online channel for example could help discussions, issues and information to stay accessible to other users too)

**Software paper**
- Paper:  "We encourage active participation in the software development process to adapt it to user needs." I would move this sentence from the paper to the documentation of the package.
- Paper: The abbreviation ICEVs is not fully introduced the first time (include the word vehicle)
- References: The two following references might be inetresting to include in the paper: "How to integrate real-world user behavior into models of the market diffusion of alternative fuels in passenger cars - An in-depth comparison of three models for Germany?" and "BIOGEME: a free package for the estimation of discrete choice models"
- References: I could not find in the JOSS guidelines whether it is fine to cite in the references your RSER paper although it is still under review.
- Dataset: Is the example dataset artificial? If not it would be nice to cite in the references where the data comes from.

I hope the suggestions are helpful to improve the package! :)
