### 2023-07-10 a04cd51
Review-response to Fabia Miorelli. Part: -Functionality-
