+++ mode_behave_public/post_analysis.py
@@ -20,15 +20,16 @@ from sklearn.cluster import KMeans, AgglomerativeClustering, MeanShift, DBSCAN
 from numba import guvectorize, prange, njit
 import seaborn as sns
 
+
 class PostAnalysis:
     """
     This class incorporates methods to analyze the estimated mixed logit
     model as well as methods to visualize the estimation & simulation results.
     """
-    
+
     def _init_(self):
-        pass       
-    
+        pass
+
     def get_AIC_MNL(self):
         """
         This method calculates the AIC of an estimated MNL-model.
@@ -40,13 +41,13 @@ class PostAnalysis:
             Akaike information criterion.
 
         """
-        
+
         LL = self.loglike_MNL()
-        
-        AIC = -2*LL[0]+2*len(self.initial_point)
-        
+
+        AIC = -2 * LL[0] + 2 * len(self.initial_point)
+
         return AIC
-    
+
     def get_BIC_MNL(self):
         """
         This method calculates the BIC of an estimated MNL-model.
@@ -58,11 +59,11 @@ class PostAnalysis:
             Bayesian information criterion.
 
         """
-        
+
         LL = self.loglike_MNL()
-        
-        BIC = -2*LL[0]+np.log(len(self.data))*len(self.initial_point)
-        
+
+        BIC = -2 * LL[0] + np.log(len(self.data)) * len(self.initial_point)
+
         return BIC
 
     def loglike_MNL(self):
@@ -83,23 +84,25 @@ class PostAnalysis:
             some numerical issue occures during calculation (e.g., log(0))
 
         """
-        
+
         top = np.zeros(shape=self.av.shape[2], dtype="float64")
         bottom = np.zeros(shape=self.av.shape[2], dtype="float64")
         for c in range(self.count_c):
             for e in range(self.count_e):
-                top += self.av[c][e] * self.choice[c][e] * np.exp(self.get_utility(c,e))
-                bottom += self.av[c][e] * np.exp(self.get_utility(c,e))
-                
-        log_res = self.weight_vector*np.log(top/bottom)
+                top += (
+                    self.av[c][e] * self.choice[c][e] * np.exp(self.get_utility(c, e))
+                )
+                bottom += self.av[c][e] * np.exp(self.get_utility(c, e))
+
+        log_res = self.weight_vector * np.log(top / bottom)
         res = np.nansum(log_res)
         number_nan = np.sum(np.isnan(log_res))
-        
+
         return res, number_nan
-    
+
     def get_utility(self, c, e):
         """
-        Calculation of the utility-function for all observations within 
+        Calculation of the utility-function for all observations within
         a given data-sample with respect to the persons choice c.
 
         Parameters
@@ -115,99 +118,121 @@ class PostAnalysis:
             Returns the utility for each observation.
 
         """
-        
+
         if c == 0:
             ASC = 0
         else:
-            ASC = self.initial_point[c-1]
-        
+            ASC = self.initial_point[c - 1]
+
         utility = (
-            ASC +
-            np.sum(
-                [self.initial_point[(self.count_c-1) + a] * 
-                 self.data[self.param['constant']['fixed'][a] + '_' + str(c) + '_' + str(e)] 
-                 for a in range(self.no_constant_fixed)], axis=0
-                ) +
-            np.sum(
+            ASC
+            + np.sum(
                 [
-                    self.initial_point[
-                        (self.count_c-1) + 
-                        self.no_constant_fixed + a
-                        ] * 
-                    self.data[self.param['constant']['random'][a] + '_' + str(c) + '_' + str(e)] 
-                    for a in range(self.no_constant_random)
-                    ], axis=0
-                ) +
-            np.sum(
-                [
-                    self.initial_point[
-                        (self.count_c-1) + 
[…]
