+++ mode_behave_public/simulation.py
@@ -15,15 +15,26 @@ applications within other models. Thus, good documentation is even more importan
 import numpy as np
 from numba import njit
 
+
 class Simulation:
     """
     This class incorporates methods to simulate previously estimated
     mixed logit and multinomial logit models.
     """
-    
-    def simulate_hh_cars(self, regiontype, hh_size,
-                         adults_working, children, htype, quali_opnv, sharing,
-                         relative_cost_per_car, age_adults, **kwargs):
+
+    def simulate_hh_cars(
+        self,
+        regiontype,
+        hh_size,
+        adults_working,
+        children,
+        htype,
+        quali_opnv,
+        sharing,
+        relative_cost_per_car,
+        age_adults,
+        **kwargs
+    ):
         """
         This method calculates the probability of a single household
         to own 0,1,2,3 or more cars.
@@ -34,20 +45,20 @@ class Simulation:
         #   adults_working: Number of working adults (replace by daily distance to work)
         #   children: Number of children in household
         #   htype: Haustyp
-        #   quali_opnv: Quality of public transport 
+        #   quali_opnv: Quality of public transport
         #   sharing: Carsharing-Membership
         #   relative_cost_per_car: Average Price of considered cars / income
         #   age_adults: Mean age of adults in the household scaled by 0.1!
-        
-        param = {'constant': 
+
+        param = {'constant':
                           {
                            'fixed':[],
                            'random':[]
                            },
                       'variable':
                           {
-                           'fixed':['urban_region', 'rural_region', 'hh_size', 'children', 
-                                    'any_car', 'htype', 'sharing', 'age_adults', 'adults_working', 
+                           'fixed':['urban_region', 'rural_region', 'hh_size', 'children',
+                                    'any_car', 'htype', 'sharing', 'age_adults', 'adults_working',
                                     'quali_opnv'],
                            'random':[relative_cost']
                            }
@@ -58,43 +69,46 @@ class Simulation:
         Probabilities, that a household owns 0,1,2,3 cars.
 
         """
-        
-        #get keyword-arguments
+
+        # get keyword-arguments
         asc_offset_hh_cars = kwargs.get("asc_offset", self.asc_offset_hh_cars)
         asc_offset = np.array(
-            [asc_offset_hh_cars['rt_' + str(regiontype)]['offset_' + str(c)] for c in range(4)]
-            )
-        
-        #define model properties below.
-        count_c = 4 #number of alternatives: 0-3
-        all_alternatives = np.array((0,1,2,3))
+            [
+                asc_offset_hh_cars["rt_" + str(regiontype)]["offset_" + str(c)]
+                for c in range(4)
+            ]
+        )
+
+        # define model properties below.
+        count_c = 4  # number of alternatives: 0-3
+        all_alternatives = np.array((0, 1, 2, 3))
         no_constant_fixed = 0
         no_constant_random = 0
         no_variable_fixed = 10
         no_variable_random = 1
-        #specify maximum number of alternatives
+        # specify maximum number of alternatives
         dim_aggr_alt_max = max(
             no_constant_fixed,
             no_constant_random,
             no_variable_fixed,
             no_variable_random,
-            )
-        
-        #Derive urban / rural regiontype according to RegioStaR2-scale from regiontype
-        if regiontype in [1,2,3,4]:
+        )
+
+        # Derive urban / rural regiontype according to RegioStaR2-scale from regiontype
+        if regiontype in [1, 2, 3, 4]:
             urban_region = 1
             rural_region = 0
-        elif regiontype in [5,6,7]:
+        elif regiontype in [5, 6, 7]:
             urban_region = 0
             rural_region = 1
         else:
             raise ValueError("Regiontype must be a value between 1-7.")
-        
[…]
