### 2023-07-10 ec4f646
Review-response to Fabia Miorelli. Part: -Functionality-
