### 2023-07-10 b0d82b0
Review-response to Fabia Miorelli. Part: -Functionality-
