### 2023-07-10 dec2d8a
Review-response to Fabia Miorelli. Part: -Functionality-
