+++ paper.bib
@@ -92,6 +92,14 @@
   url = "https://biogeme.epfl.ch/documents.html"
 }
 
+@proceedings{Bierlaire2003,
+  title = "BIOGEME: a free package for the estimation of discrete choice models",
+  year = "2003",
+  booktitle = "3rd Swiss Transport Research Conference",
+  venue = "Monte Verita, Ascona",
+  month = "March",
+}
+
 @misc{Brathwaite2020,
   author = "T. Brathwaite",
   title = "pylogit",
@@ -129,6 +137,23 @@
 @article{Reul2022,
   author   = "J. Reul and T. Grube and J. Linßen and D. Stolten",
   title    = "Modeling Behavioral Change in Transport Futures",
-  journal  = "under Review in Renewable and Sustainable Energy Reviews",
+  journal  = "SSRN preprint",
   year     = 2022,
+  doi = "http://dx.doi.org/10.2139/ssrn.4230060"
+}
+
+@article{Reul2023,
+  author   = "J. Reul and A. Chapman and T. Grube and J. Linßen and D. Stolten",
+  title    = "A Simulation of the Market Diffusion of Electric Passenger Cars in Japan and Germany",
+  journal  = "Zenodo preprint",
+  year     = 2023,
+  doi = "10.5281/zenodo.7742056"
+}
+
+@article{Gnann2022,
+  author   = "T. Gnann and D. Speth and K. Seddig and M. Stich and W. Schade and J.J. Gómez Vilchez",
+  title    = "How to integrate real-world user behavior into models of the market diffusion of alternative fuels in passenger cars - An in-depth comparison of three models for Germany",
+  journal  = "Renewable and Sustainable Energy Reviews",
+  year     = 2022,
+  doi = "10.1016/j.rser.2022.112103"
 }
