+++ README.rst
@@ -22,13 +22,19 @@ the divergent choice behavior of different individuals or consumer groups can
 only be studied to a limited degree. Mixed logit models overcome this deficiency and 
 allow for the analysis of preference distributions across base populations.
 
-Use MO|DE.behave via web app: https://julianreul-streamlit-mode-behave-home-nmwcdr.streamlit.app/
+Communication and contribution:
+We encourage active participation in the software development process to adapt 
+it to user needs. If you would like to contribute to the project or report any bugs, 
+please refer to the contribution-file or simply create an issue in the repository.
+For any other interests (e.g. potential research collaborations), please 
+directly contact the project maintainers via email, as indicated and 
+updated on GitHub.
 
 Installation
 ============
 1. Download or clone the repository to a local folder.
 #. Open (Anaconda) Prompt.
-#. Create a new environment from reference_environment.yml file::
+#. Create a new environment from reference_environment.yml file (recommended)::
 
       conda env create -f reference_environment.yml
       
@@ -49,18 +55,52 @@ Installation
 
 Workflow
 ========
-1. Import model with::
 
+This section explains an exemplary workflow from model setup to estimation 
+and post-processing for a sub-sample of survey data on household decisions
+on the type of propulsion technology when purchasing a new car.
+The propulsion types are differentiated into "ICEV: Internal combustion engine vehicle",
+"PHEV: Plug-in Hybrid Electric Vehicle", "BEV: Battery Electric Vehicle", 
+and "FCEV: Fuel Cell Electric Vehicle".
+The data was collected in the year 2021 among German households and 
+the respective sub-sample is provided with the model (./mode_behave_public/InputData/example_data.csv).
+The complete script accessible as well (./mode_behave_public/Deployments/example_estimation.py)
+
+1. Import model and required modules with::
+
+      import numpy as np
+      import pandas as pd
+      
       import mode_behave_public as mb
       
-2. Load data with (exemplary import requires pandas-module)::
-    
+2. Load data with (PATH_TO_DATA requires individual definition. See below for further documentation on required data formats.)::
+      
       example_data = pd.read_csv(PATH_TO_DATA + "example_data.csv")
+      
+3. Definition of model parameters "PURCHASE_PRICE", "RANGE", and "CHARGING_FUELING_TIME". See section "Structure of Parameters and Input Data" for further information::
+      
+      param_fixed = []
+      param_random = ['PURCHASE_PRICE', 'RANGE', 'CHARGING_FUELING_TIME']
 
-2. Initialize a model with::
+      param_temp = {'constant': 
+                        {
+                         'fixed':[],
+                         'random':[]
+                         },
+                    'variable':
+                        {
+                         'fixed': [],
+                         'random':[]
+                         }
+                    }
+
+      param_temp['variable']['fixed'] = param_fixed
+      param_temp['variable']['random'] = param_random  
+
+4. Initialize a model with::
     
       model = mb.Core(
-          param = param_temp, 
+          param=param_temp, 
           data_in=example_data, 
           alt=4,
           equal_alt=1,
@@ -69,9 +109,18 @@ Workflow
       
    The structure of the input data and the parameter-input are given below.
 
-3. Estimate the model with::
+5. Estimate the model with::
 
-    model.estimate_mixed_logit(**kwargs)  
+    model.estimate_mixed_logit(
+        min_iter=10, 
+        max_iter=1000,
+        tol=0.01,
+        space_method = 'std_value',
+        scale_space = 2,
+        max_shares = 1000,
+        bits_64=True,
+        t_stats_out=False
+        )
       
     The estimation of the mixed logit model can be modified by definition of keyword-arguments
     during instantiation and within the estimation-method itself.
@@ -114,17 +163,64 @@ Workflow
[…]
