+++ .github/workflows/testing.yml
@@ -6,7 +6,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ["3.8"]
+        python-version: ["3.8.17"]
 
     steps:
       - uses: actions/checkout@v3

[…]
