+++ mode_behave_public/post_analysis.py
@@ -1,9 +1,4 @@
-# -*- coding: utf-8 -*-
 """
-Created on Fri Jan 29 16:39:30 2021
-
-@author: j.reul
-
 This module holds the class "PostAnalysis", which incorporates functionality
 to analyze the estimated mixed logit model.
 """
@@ -73,7 +68,6 @@ class PostAnalysis:
 
         Parameters
         ----------
-        A "point" with all coefficients of MLN-attributes.
 
         Returns
         -------
@@ -197,10 +191,15 @@ class PostAnalysis:
 
         For reference on log-likelihood calculation see:
             Ch. 5.5 (pp. 118) in "Discrete Choice Analysis", by Ben-Akiva (1985)
+            
+        Parameters
+        ----------
+        points_in : array, optional
+            An array of preference points, if divergent to previously estimated points.
 
         Returns
         -------
-        Float64
+        log-likelihood : float64
             Returns the log-likelihood (LL) of the estimated mixed logit model.
 
         """
@@ -324,22 +323,21 @@ class PostAnalysis:
         attribute : str
             Name of the attribute to be visualized.
 
-        kwargs save_fig_path : string
+        save_fig_path : string, optional
             Path, which indicated the place where to store the visualization
             as a .png-file.
 
-        kwargs names_choice_options : dict
+        names_choice_options : dict, optional
             If given, this shall be a dictionary, which holds the
             names of the choice options as values and the numerical
             indication of the choice option (0,1,2,...) as keys.
 
-        kwargs shift_t_stats : float
+        shift_t_stats : float, optional
             Shifts the text-fields which indicate the value of the t-statistic
             in positive or negative direction..
 
         Returns
         -------
-        None.
 
         """
         # get keyword arguments
@@ -406,46 +404,37 @@ class PostAnalysis:
 
         Parameters
         ----------
-        kwargs return_res : Boolean
+        return_res : Boolean, optional
             If True, the clustering results are returned. Defaults to False.
-
-        kwargs return_figure : Boolean
+        return_figure : Boolean, optional
             If True, the rendered figure is returned. Defaults to False.
-
-        kwargs cluster_method : string
+        cluster_method : string, optional
             Specification of the clustering method, which shall be used to cluster
             the preferences estimated by the mixed logit model.
             Defaults to "kmeans". Other options: "agglo", "meanshift", "dbscan".
-
-        kwargs scale_individual : Boolean
+        scale_individual : Boolean, optional
             If True, the x- and y-axis of the visualizations are scaled to one.
             This eases the comparison of the different attributes. The scale
             is indicates on the respective axes and is very important for
             quantitative and qualitative interpretations. Defaults to False.
-
-        kwargs external_points : array
+        external_points : array, optional
             This array holds further parameter points in the parameter space,
             which should be visualized as reference points. E.g.: The initial
             point, as being calculated by the multinomial logit model.
-
-        kwargs k : int
+        k : int, optional
             Number of cluster centers to be calculated. Defaults to 3.
-
-        kwargs save_fig_path : string
+        save_fig_path : string, optional
             Path, which indicated the place where to store the visualization
             as a .png-file.
-
-        kwargs name_scenario : string
+        name_scenario : string, optional
             The scenario name can be added additionally, to distinguish
             several scenarios.
-
-        kwargs bw_adjust : float
+        bw_adjust : float, optional
             This value adjusts the smoothing of the visualized preference
             distribution. A higher value increases the smoothing of the
             displayed curve, but may conceal certain findings in the distribution.
             Defaults to 0.03.
-
-        kwargs names_choice_options : dict
+        names_choice_options : dict, optional
             If given, this shall be a dictionary, which holds the
[…]
