+++ mode_behave_public/simulation.py
@@ -1,9 +1,4 @@
-# -*- coding: utf-8 -*-
 """
-Created on Fri Jan 29 16:39:30 2021
-
-@author: j.reul
-
 This module holds the class "Simulation", which incorporates functionality
 to simulate the probabilities with stored parameters from previous estimations.
 
@@ -40,33 +35,31 @@ class Simulation:
         to own 0,1,2,3 or more cars.
         Parameters
         ----------
-        #   regiontype: RegioStaR7-classification of regions
-        #   hh_size: Household-size
-        #   adults_working: Number of working adults (replace by daily distance to work)
-        #   children: Number of children in household
-        #   htype: Haustyp
-        #   quali_opnv: Quality of public transport
-        #   sharing: Carsharing-Membership
-        #   relative_cost_per_car: Average Price of considered cars / income
-        #   age_adults: Mean age of adults in the household scaled by 0.1!
-
-        param = {'constant':
-                          {
-                           'fixed':[],
-                           'random':[]
-                           },
-                      'variable':
-                          {
-                           'fixed':['urban_region', 'rural_region', 'hh_size', 'children',
-                                    'any_car', 'htype', 'sharing', 'age_adults', 'adults_working',
-                                    'quali_opnv'],
-                           'random':[relative_cost']
-                           }
-                      }
+        regiontype: int
+            RegioStaR7-classification of regions (1-7), where 1-4 are urban 
+            classifications and 5-7 rural classifications. 
+        hh_size: int
+            Household size
+        adults_working: int 
+            Number of working adults
+        children: int
+            Number of children in household
+        htype: int
+            1, if individual house, 0, if multi-apartment house
+        quali_opnv: int
+            Quality of public transport on a scale from 1-4 
+            (1: Very Bad, 2: Bad, 3: Good, 4: Very Good)
+        sharing: float
+            Carsharing membership (1: Yes, 0: no)
+        relative_cost_per_car: 
+            Average Price of considered cars / net household income
+        age_adults: float
+            Mean age of adults in the household scaled by 0.1!
 
         Returns
         -------
-        Probabilities, that a household owns 0,1,2,3 cars.
+        res_temp : array
+            Probabilities, that a household owns 0,1,2, or 3 cars.
 
         """
 
@@ -234,7 +227,8 @@ class Simulation:
         distance : int
             Travel distance.
         regiontype : int
-            Region of travel.
+            RegioStaR7-classification of regions (1-7), where 1-4 are urban 
+            classifications and 5-7 rural classifications. 
 
         Returns
         -------
@@ -271,7 +265,8 @@ class Simulation:
         distance : int
             Travel distance.
         regiontype : int
-            Region of travel.
+            RegioStaR7-classification of regions (1-7), where 1-4 are urban 
+            classifications and 5-7 rural classifications. 
 
         Returns
         -------
@@ -296,7 +291,8 @@ class Simulation:
         distance : int
             Travel distance.
         regiontype : int
-            Region of travel.
+            RegioStaR7-classification of regions (1-7), where 1-4 are urban 
+            classifications and 5-7 rural classifications. 
 
         Returns
         -------
@@ -319,26 +315,26 @@ class Simulation:
     ):
         """
         This method calculates the probability, that a transport mode is chosen.
+        
         Parameters
[…]
