+++ README.rst
@@ -30,6 +30,7 @@ For any other interests (e.g. potential research collaborations), please
 directly contact the project maintainers via email, as indicated and 
 updated on GitHub.
 
+
 Installation
 ============
 1. Download or clone the repository to a local folder.

[…]
