+++ .readthedocs.yml
@@ -7,8 +7,6 @@ version: 2
 # Set the version of Python and other tools you might need
 build:
   os: ubuntu-22.04
-  tools:
-    python: "3.10"
 
 mkdocs:
   configuration: mkdocs.yml
