+++ .readthedocs.yml
@@ -7,6 +7,13 @@ version: 2
 # Set the version of Python and other tools you might need
 build:
   os: ubuntu-22.04
+  tools:
+    python: "3.11"
 
 mkdocs:
   configuration: mkdocs.yml
+
+# Optionally declare the Python requirements required to build your docs
+python:
+  install:
+  - requirements: requirements.txt
