### 2023-07-11 #5265 @julianreul
Hi @fabmio,

thanks for your detailed review and great remarks! I have adressed all of your comments in recent commits. Please find a summary of the changes (**bold**) in response to your remarks (*italic*) below:

**General checks**

- *Documentation: Overall I think it would be really nice and easier to navigate to have a separate software documentation page (e.g. readthedocs) than the readme alone. This way it would also be possible to look up through the docstrings all methods in each class and their respective arguments from the documentation instead of looking into the codebase.* --> **I created an additional sofware-documentation using GitHub Pages based on the docstrings of the modules, classes and methods: https://fzj-iek3-vsa.github.io/mode_behave/**

**Functionality**

- *Code: There are lots of trailing whitespaces and white lines containing white spaces in the code as well as missing white spaces after block comments, missing white spaces after commas in arguments and continuation lines which are in some cases over and underindented. I would suggest to install an autoformatter and riformat the entire codebase.* --> **I used the Python code formater black to reformat the entire code base according to PEP 8 standards**
- *Code: In the estimation file on line 114 avoid the usage of a bare Except and on line 386 rename the variable l which is ambiguous (is it an I or small L?)* --> **1) I modified the try-except formulation in the estimation-module to only catch the desired Attribute-Error. 2) In this case, I would like to stick to the declaration of the iteration variable l due to several reasons: a) Within a code editor, such as VS Code, the small letter l is not ambiguous and it is even recommended to use single letters for looping. b) I am using this letter l to loop over the length of the data input in several places in the code base. Thus, I'd have to change the code in many places, which could provoke typo-errors during this process. c) The code, which you are referring to is at the very bottom of the code base. Thus, changes (and potential typo-errors) could propogate upwards, endangering the functionality of the whole package.**
- *Testing: I run both tests from terminal: the test_simulation ran without errors. In the test_estimation one test failed with the following error message: AssertionError: False is not true on line 252 in test_estimate_mixed_logit (self.assertTrue(np.allclose(np.array(model.points), points_compare, atol=0.1))* --> **This error message occurred due to differences in the used scipy-package. The reference_environment.yaml, the setup.py and the requirements.txt specified the required scipy-version with scipy>=1.10.1, which was ambiguous. I changed that definition to scipy==1.10.1. Now, the error shouldn't occur anymore, after pulling the latest changes from the repo.**

**Documentation**

- *Example usage: I would rewrite the example usage in the documentation to reflect 1:1 the example_estimation.py file so that can work as a guideline and introduction to the tool for new users (explicitly adding for example the pandas import)* --> **I adapted the readme-file according to your suggestions.**
- *Contribution file: In the contribution file I would add that the new contributions should also be properly documented.* --> **I adapted the contribution-file according to your suggestions.**
- *Visualisations: it would be nice to have also the visualisation module documented.* --> **There is no visualization-module to be documented. Visualization-methods are integrated in the post_analysis module. The important visualization methods are referenced in the example_estimation.py script and on GitHub Pages**
- *Issue reporting: I would indicate in the documentation where users could report issues or problems with the software (issue board, email, online channel like gitter, slack, or similar? an online channel for example could help discussions, issues and information to stay accessible to other users too)* --> **Apart from the contribution-file I added a new paragraph in the readme, called "communication and contribution", which explains the procedures of how to interact and contribute. For issue-reporting I would like to stick to the regular workflow via GitHub-issues as an established procedure in software development and to avoid parallel threads.**

**Software paper**

- *Paper: "We encourage active participation in the software development process to adapt it to user needs." I would move this sentence from the paper to the documentation of the package.* --> **I added this sentence to the readme**
- *Paper: The abbreviation ICEVs is not fully introduced the first time (include the word vehicle)* --> **I adapted the paper accordingly.**
- *References: The two following references might be inetresting to include in the paper: "How to integrate real-world user behavior into models of the market diffusion of alternative fuels in passenger cars - An in-depth comparison of three models for Germany?" and "BIOGEME: a free package for the estimation of discrete choice models"* --> **I included both references in the paper. Thanks especially for the first paper, which is really interesting for me personally and my current research.**
- *References: I could not find in the JOSS guidelines whether it is fine to cite in the references your RSER paper although it is still under review.* --> **I changed the reference to point to the preprint of this paper, which also has a constant identifier (DOI)**
- *Dataset: Is the example dataset artificial? If not it would be nice to cite in the references where the data comes from.* --> **The dataset is not artificial. It's a sub sample from survey data collected among German households in 2021. Unfortunately, we cannot share the whole dataset due to ongoing publication processes. However, we can share this sub-sample, which is used in the example. I added some information on the collection of the survey data in the paper.**

Thanks again for your remarks and your willingness to review this piece of software. I hope to have addressed your comments sufficiently.

Best
Julian
