### 2023-08-01 #5265 @fabmio
> @fabmio and @abhiramm7 can you please comment on whether the author has adequately addressed your concerns? If you can complete the checklist or verbally sign off, that will let us proceed to the next stage of editing. If you have unaddressed concerns, please let @julianreul know.

@sbenthall: All my comments have been addressed as well. I disagree with the arguments to the second point related to the code functionality. Nevertheless I deem it fine to continue with the publication process as the mentioned iterators are used in a limited scope and are not detrimental towards readability and maintainability of the code.

@julianreul: thank you for addressing all the comments!
