### 2023-08-07 #5265 @sbenthall
Great!

@julianreul We can proceed to the next step of the process.

These are:

- Final proofreading. Last chance to make edits!
- Tagging a release, and telling me the release number
- Making an archive of the release on Zenodo, and telling me the DOI
