### 2023-08-08 31ccfb3
Updated references in paper.md
