### 2023-08-08 d4d0893
Updated references in paper.md
