### 2023-08-08 #5265 @julianreul
Thank you @abhiramm7 and @fabmio for your reviews and @sbenthall for handling this submission up to this point.

I conducted a final proof read, tagged a release and created an archive on Zenodo:
- Release:    [v1.0.12] (https://github.com/FZJ-IEK3-VSA/mode_behave/releases/tag/v1.0.12)
- Zenodo:   10.5281/zenodo.8223829
