### 2023-08-08 #5265 @sbenthall
Uh oh @julianreul

Looks like the DOIs for the references are throwing some errors.

I'll need you to update these.

I'll then run the check-references script to make sure it clears.

Once we do all this, we're going to have to go through the tagging-and-archiving dance again.
