### 2023-08-08 d358e41
Updated DOI of a reference in paper.bib
