+++ paper.bib
@@ -139,7 +139,7 @@
   title    = "Modeling Behavioral Change in Transport Futures",
   journal  = "SSRN preprint",
   year     = 2022,
-  doi = "http://dx.doi.org/10.2139/ssrn.4230060"
+  doi = "10.2139/ssrn.4230060"
 }
 
 @article{Reul2023,
