### 2023-08-08 #5265 @julianreul
Dear @sbenthall, I have changed the incorrect reference in the paper.bib and created a new release on GitHub (just to be on the safe side), which comes with a new DOI:
- Release: [v1.0.13] https://github.com/FZJ-IEK3-VSA/mode_behave/releases/tag/v1.0.13
- Zenodo: 10.5281/zenodo.8225871
